Bugfix?
Trying to fix custom recursive copy so fs-extra can be dumped.
